backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==0.28.1
faiss-cpu==1.7.4
numpy==1.24.3
requests==2.31.0
aiofiles==0.23.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4